[ADD] Routing, and empty Login and Home screens (#8)
